+++ include/MeshKernel/Orthogonalizer.hpp
@@ -50,7 +50,7 @@ namespace meshkernel
         /// @brief node
         /// @brief connectedNode
         /// @returns The contribution of orthogonalizer to the left hand side, the linear system
-        [[nodiscard]] double GetWeight(size_t node, int connectedNode)
+        [[nodiscard]] double GetWeight(size_t node, size_t connectedNode)
         {
             return m_weights[node][connectedNode];
         }
@@ -59,7 +59,7 @@ namespace meshkernel
         /// @brief node
         /// @brief connectedNode
         /// @returns The contribution of orthogonalizer to the right hand size
-        [[nodiscard]] double GetRightHandSide(size_t node, int connectedNode)
+        [[nodiscard]] double GetRightHandSide(size_t node, size_t connectedNode)
         {
             return m_rhs[node][connectedNode];
         }
